+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [Unreleased]
+
+### Changed
+
+- Changed field `astro_addr: Addr` to `astro_token: AssetInfo` on struct `AstroportStaking`.
+  - This is a breaking change.
+  - This is to support chains where ASTRO is a native token.
+
 # [0.3.1] - 2023-07-20
 
 ### Changed